use note frequencies rather than a sustained note to lock in note on headstock
