+++ src/Components/CustomDropdown/CustomDropdown.scss
@@ -6,6 +6,10 @@
     font-family: inherit;
     font-size: 14px;
     color: var(--color-text-primary);
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    
   }
   
   .dropdown-box {
@@ -17,8 +21,10 @@
     border: 1px solid var(--color-border);
     border-radius: 4px;
     cursor: pointer;
+    -webkit-user-select: none;
     user-select: none;
     transition: border 0.2s ease;
+    margin-left: 1rem;
   }
   
   .dropdown-box:hover {
@@ -47,8 +53,8 @@
   }
   
   .option:hover {
-    // background-color: var(--color-surface);
-    background-color: red;
+     background-color: greenyellow;
+    
   }
   
   .selected-option {
