--- src/Components/Headstock.scss
@@ -1,11 +1,0 @@
-.max-w-full {
-   
-    padding: 0rem;
-   margin: 0rem;
-   
-
-}
-
-.arrow-container {
-    
-}
