+++ src/Headstock.scss
@@ -1,0 +1,11 @@
+// .max-w-full {
+   
+//     padding: 0rem;
+//    margin: 0rem;
+   
+
+// }
+
+// .arrow-container {
+    
+// }
